requirements.txt
python-dotenv
requests
beautifulsoup4
PyMuPDF
numpy
openai
click
rich
python-docx
